refactor: move DefaultStore implementation in block producer into store module
